Last changes

+++ src/menus/HighScoreSong2Button.java
@@ -48,7 +48,7 @@ public class HighScoreSong2Button extends TexturedMenuButton {
         if (percent == 0) {
             return "";
         } else {
-            return percent + "";
+            return (int)percent + "";
         }
     }
